Refactor store module layout and exports
Move concrete store implementations into a new stores::impls module and
introduce a stores::config module for configuration types. Update
internal use paths and public re-exports accordingly. Add publish_compat
and validation_compat shims to preserve the previous external API and
adjust CLI imports and manager references to the new layout.
