+++ crates/store/src/publish_compat.rs
@@ -1,0 +1,6 @@
+//! Publish compatibility module for backward compatibility
+//!
+//! This module re-exports all publish-related types from the manager module
+//! to maintain backward compatibility with external crates.
+
+pub use crate::manager::publish::*;
