+++ crates/store/src/stores/config/registry.rs
