+++ crates/store/src/validation_compat.rs
@@ -1,0 +1,6 @@
+//! Validation compatibility module for backward compatibility
+//!
+//! This module re-exports all validation-related types from the registry module
+//! to maintain backward compatibility with external crates.
+
+pub use crate::registry::validation::*;
